+++ projects/egeo/src/lib/st-table/shared/table-header.interface.ts
@@ -8,8 +8,24 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
+
+export interface StFilterElement {
+   id: String | number;
+   name: String | number;
+   selected?: boolean;
+}
+
+export interface StFilterHeader {
+   filterConfig?: StFilterElement[];
+   showSettingBtn?: boolean;
+   title: string;
+   buttonText?: string;
+}
+
 export interface StTableHeader {
    id: string;
    label: string;
    sortable?: boolean;
+   filters?: StFilterHeader;
+   filterable?: boolean;
 }